add option to override specific theme options
added themeoverrides.txt to allow the user to override FONT_LIST_PAD_RIGHT and CONTENT_WIDTH of a theme.  This will be useful for users that are displaying boxart and do not want the list item text to get covered by the box art.  More settings can be added later if a need arises.